requirements.txt
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
distlib==0.3.9
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
fonttools==4.55.3
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
lizard==1.17.13
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-venn==1.1.1
mysql-connector-python==9.2.0
numpy==2.2.1
packaging==24.2
pillow==11.1.0
platformdirs==4.3.7
PyDriller==2.7
Pygments==2.18.0
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.2
requests==2.32.3
scipy==1.14.1
six==1.17.0
smmap==5.0.2
types-pytz==2024.2.0.20241221
urllib3==2.4.0
virtualenv==20.30.0
Werkzeug==3.1.3




